requirements.txt
# Lily-Discord-Adapter dependencies
discord.py>=2.3.0
httpx>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
aiofiles>=23.0.0
PyNaCl>=1.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
yt-dlp>=2024.4.9
